pyproject.toml
[project]
name = "ml-coarsening"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hydra-core>=1.3.2",
    "lightning>=2.5.1.post0",
    "mlflow>=2.22.0",
    "torch>=2.7.0",
]
